+++ src/lib/util/globalInteractions.ts
@@ -81,14 +81,6 @@ export function makeOpenSeedPackButton() {
 		.setEmoji('977410792754413668');
 }
 
-export function makeAutoFarmButton() {
-	return new ButtonBuilder()
-		.setCustomId('AUTO_FARM')
-		.setLabel('Auto Farm')
-		.setStyle(ButtonStyle.Secondary)
-		.setEmoji('630911040355565599');
-}
-
 export function makeAutoContractButton() {
 	return new ButtonBuilder()
 		.setCustomId('AUTO_FARMING_CONTRACT')
