+++ src/lib/util/smallUtils.ts
@@ -152,3 +152,11 @@ export function makeEasierFarmingContractButton() {
 		.setStyle(ButtonStyle.Secondary)
 		.setEmoji('977410792754413668');
 }
+
+export function makeAutoFarmButton() {
+	return new ButtonBuilder()
+		.setCustomId('AUTO_FARM')
+		.setLabel('Auto Farm')
+		.setStyle(ButtonStyle.Secondary)
+		.setEmoji('630911040355565599');
+}
